🎨 Convert to Navbar layout
